# Implement points array backend