--- a/sleap_io/__init__.py
+++ b/sleap_io/__init__.py
@@ -5,8 +5,6 @@ from sleap_io.model.camera import Camera, CameraGroup
 from sleap_io.model.skeleton import Node, Edge, Skeleton, Symmetry
 from sleap_io.model.video import Video
 from sleap_io.model.instance import (
-    Point,
-    PredictedPoint,
     Track,
     Instance,
     PredictedInstance,