--- a/tests/io/test_slp.py
+++ b/tests/io/test_slp.py
@@ -8,8 +8,6 @@ from sleap_io import (
     Instance,
     LabeledFrame,
     Track,
-    Point,
-    PredictedPoint,
     PredictedInstance,
     Labels,
     SuggestionFrame,