… sleap_io/model/skeleton.py
… tests/model/test_instance.py
… tests/model/test_labeled_frame.py