This PR implements point storage using structured numpy arrays (similar to core SLEAP).

This provides a huge speed up when loading large datasets which can easily scale to millions or billions of points. Before this PR, the overhead of creating this many `Point` and `PredictedPoint` objects would dominate runtime. By creating fewer numpy arrays with preallocated memory, loading is  super fast.

This does introduce breaking changes to anything that used the `Point` objects directly, but otherwise, almost all of the API is identical to before.

## Benchmark

### Script
```py
import sleap_io as sio
import cProfile
import pstats
import io

if __name__ == "__main__":
    profiler = cProfile.Profile()
    profiler.enable()
    labels = sio.load_slp(
        "/Users/talmo/Downloads/cohort2.220503_153813.Camera0.mov.00000.slp"
    )
    profiler.disable()

    # Save and print results
    stream = io.StringIO()
    stats = pstats.Stats(profiler, stream=stream)
    stats.strip_dirs()
    stats.sort_stats(pstats.SortKey.TIME)
    stats.print_stats()
    print(stream.getvalue())
```

### Data

```
(sleap-io) talmo@Talmos-MBP-M2 sleap-io % python -c "import sleap_io; print(sleap_io.load_slp('/Users/talmo/Downloads/cohort2.220503_153813.Camera0.mov.00000.slp'))"
Labels(labeled_frames=42500, videos=1, skeletons=1, tracks=2, suggestions=0)

(sleap-io) talmo@Talmos-MBP-M2 sleap-io % python -c "import sleap_io; print(sleap_io.load_slp('/Users/talmo/Downloads/cohort2.220503_153813.Camera0.mov.00000.slp').numpy().shape)"
(42500, 2, 15, 2)
```

**Total:** 1,275,000 points

### Before (4.7 secs)

```
(sleap-io) talmo@Talmos-MBP-M2 sleap-io % python load_slp_profile.py
         2455609 function calls (2371009 primitive calls) in 4.702 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    3.074    3.074    3.154    3.154 slp.py:800(read_pred_points)
    84593    0.521    0.000    0.584    0.000 instance.py:211(_convert_points)
        1    0.518    0.518    1.245    1.245 slp.py:816(read_instances)
    42500    0.172    0.000    0.172    0.000 <attrs generated init sleap_io.model.labeled_frame.LabeledFrame>:1(__init__)
        1    0.097    0.097    4.696    4.696 slp.py:1041(read_labels)
    84593    0.081    0.000    0.720    0.000 <attrs generated init sleap_io.model.instance.PredictedInstance>:1(__init__)
  1268895    0.066    0.000    0.066    0.000 <attrs generated init sleap_io.model.instance.PredictedPoint>:1(__init__)
    84593    0.037    0.000    0.622    0.000 instance.py:252(__attrs_post_init__)
253821/169228    0.024    0.000    0.035    0.000 {built-in method builtins.len}
        7    0.020    0.003    0.022    0.003 dataset.py:765(__getitem__)
   169186    0.019    0.000    0.019    0.000 {method 'keys' of 'dict' objects}
    84593    0.012    0.000    0.017    0.000 validators.py:94(__call__)
    84593    0.011    0.000    0.015    0.000 skeleton.py:306(__len__)
    84594    0.010    0.000    0.010    0.000 {method 'values' of 'dict' objects}
        1    0.008    0.008    0.015    0.015 labels.py:64(update)
        1    0.006    0.006    4.702    4.702 main.py:11(load_slp)
```

**Runtime:** 4.702
**Points per sec:** 271161.208

### After (1.7 secs)

```
(sleap-io) talmo@Talmos-MBP-M2 sleap-io % python load_slp_profile.py 
         2201830 function calls (2032637 primitive calls) in 1.696 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.625    0.625    1.462    1.462 slp.py:808(read_instances)
    84593    0.283    0.000    0.283    0.000 {built-in method numpy.empty}
    84593    0.183    0.000    0.242    0.000 _shape_base_impl.py:621(column_stack)
    84593    0.100    0.000    0.550    0.000 instance.py:456(empty)
        1    0.093    0.093    1.696    1.696 slp.py:1051(read_labels)
       10    0.080    0.008    0.080    0.008 files.py:215(make_fid)
423007/253821    0.047    0.000    0.069    0.000 {built-in method builtins.len}
    84593    0.038    0.000    0.074    0.000 instance.py:496(_convert_points)
   169186    0.025    0.000    0.025    0.000 {built-in method numpy.array}
    84595    0.025    0.000    0.025    0.000 skeleton.py:242(node_names)
    42500    0.022    0.000    0.022    0.000 <attrs generated init sleap_io.model.labeled_frame.LabeledFrame>:1(__init__)
   169186    0.022    0.000    0.032    0.000 skeleton.py:307(__len__)
   211747    0.020    0.000    0.020    0.000 {method 'append' of 'list' objects}
    84593    0.019    0.000    0.093    0.000 instance.py:233(__attrs_post_init__)
    84593    0.017    0.000    0.026    0.000 shape_base.py:208(_arrays_for_stack_dispatcher)
        7    0.016    0.002    0.019    0.003 dataset.py:765(__getitem__)
    84593    0.016    0.000    0.109    0.000 <attrs generated init sleap_io.model.instance.PredictedInstance>:1(__init__)
    84593    0.015    0.000    0.041    0.000 _shape_base_impl.py:617(_column_stack_dispatcher)
   169186    0.013    0.000    0.013    0.000 {built-in method numpy.asanyarray}
    84652    0.009    0.000    0.009    0.000 {built-in method builtins.hasattr}
        1    0.008    0.008    0.015    0.015 labels.py:64(update)
    84593    0.005    0.000    0.005    0.000 multiarray.py:180(concatenate)
    42500    0.005    0.000    0.007    0.000 labeled_frame.py:41(__iter__)
    42500    0.002    0.000    0.002    0.000 {built-in method builtins.iter}
        2    0.001    0.000    0.001    0.000 {method 'read' of '_io.BufferedReader' objects}
        1    0.001    0.001    1.696    1.696 main.py:11(load_slp)
```

**Runtime:** 1.696
**Points per sec:** 751768.868
**Speed-up:** 2.77240566x